create working dev environment
When pulling from the repo, I noticed that ‘grunt dev’ and ‘grunt
build’ failed. The main issue was that during the Browserify task,
Grunt was looking for a nonexistant bower_components directory.

This change appends app/ to the bower_components directory, giving a
valid file path.

Grunt also was unable to start the Express server as Express was not
listed as an npm dependency. This commit appends that, too. I do not
know which version is most desirable, however, and so just fetch the
latest version.

Now, after a git clone of the repo, grunt dev works flawlessly.