Convert grunt-browserify "external" to "exclude"
/cc @jmreidy

Don't know why this was necessary. I deployed last night to prod and everything was fine—compiled today and there was different output. I compared the node_module versions of grunt-browserify, browserify, browser-resolve, and resolve, but couldn't a difference.

On prod, passing 'buckets' to external resulted in "buckets: false" in the mapping/output. Today, it was assigned an ID. Moving to exclude nixes it from the mapping altogether which works...

Realize my setup is somewhat elaborate/complex (trying to "share" a module between two builds), but wish I understood why this seemed to change overnight.