Design Builds, Staging preview, Harp, Dropbox WIP
Sincere apologies in advance for the monstrous commit. There were a lot of interconnected parts to this, and I didn't want to ship without at least a working local demo and a healthy group of tests.

I'll be doing more docs and a blog post soon, but here are the features added:
* Automatic staging version of the site at staging.yoursite.com
* Snapshots of theme builds, with rollback support
* Deploy design updates via Git/FTP (or soon Dropbox)
* Built in preprocessor support for Jade, CoffeeScript, Sass, Stylus, LESS, etc. for template/design assets (via Harp.js)
* Template editor can now edit most client files, like CSS, JS, and preprocessed files, in addition to Handlebars templates
* Templates/statics still served from local filesystem (fast)
* Use a symlink for the live build so updates are instant
* Editor files use a tree view for paths in names, command-enter now saves files in the Design Editor

While this sounds like overkill (and likely is, a bit), these features work together to provide a web design environment unlike any other CMS. The UI for all of these features is still light and straight-forward.