Added a pre-installation redirection to /admin
Instead of a 404 error Buckets will now redirect
the user to the admin dashboard on initial launch.

This approach uses the routes array to determine,
whether the app is in a pre-install state.

Performance wise does this commit change nothing, since
the if check has just been moved to another location.

A 302 'Found' code is sent along with the redirect (none permanent).